feat: add dishes module
