+++ src/dishes/entity/dish.entity.ts
@@ -1,0 +1,9 @@
+export class User {
+    constructor(
+        public id: string,
+        public name: string,
+        public description: string,
+        public price: string,
+        public rating: number,
+    ) {}
+}
